moved `graph`, `alloc` and `process` into `renderer`
