+++ src/lib.rs
@@ -49,14 +49,13 @@ pub const RENDER_QUANTUM_SIZE: usize = 128;
 /// Maximum number of channels for audio processing
 pub const MAX_CHANNELS: usize = 32;
 
-pub mod alloc;
 pub mod buffer;
 pub mod context;
 pub mod control;
 pub mod media;
 pub mod node;
 pub mod param;
-pub mod process;
+pub mod renderer;
 pub mod spatial;
 
 #[cfg(test)]
@@ -66,7 +65,6 @@ mod snapshot;
 mod io;
 
 mod analysis;
-mod graph;
 mod message;
 
 /// Number of samples processed per second (Hertz) for a single channel of audio
