+++ src/param.rs
@@ -2,12 +2,11 @@
 use std::slice::{Iter, IterMut};
 use std::sync::Arc;
 
-use crate::alloc::AudioRenderQuantum;
 use crate::context::AudioContextRegistration;
 use crate::node::{
     AudioNode, ChannelConfig, ChannelConfigOptions, ChannelCountMode, ChannelInterpretation,
 };
-use crate::process::{AudioParamValues, AudioProcessor};
+use crate::renderer::{AudioParamValues, AudioProcessor, AudioRenderQuantum};
 use crate::{AtomicF32, SampleRate, RENDER_QUANTUM_SIZE};
 
 use crossbeam_channel::{Receiver, Sender};
