+++ src/spatial.rs
@@ -1,14 +1,12 @@
 //! Spatialization/Panning primitives
 //!
 //! Required for panning algorithm, distance and cone effects of [`crate::node::PannerNode`]s
-
-use crate::alloc::AudioRenderQuantum;
 use crate::context::{AsBaseAudioContext, AudioContextRegistration, AudioParamId};
 use crate::node::{
     AudioNode, ChannelConfig, ChannelConfigOptions, ChannelCountMode, ChannelInterpretation,
 };
 use crate::param::{AudioParam, AudioParamOptions, AudioParamRaw, AutomationRate};
-use crate::process::{AudioParamValues, AudioProcessor};
+use crate::renderer::{AudioParamValues, AudioProcessor, AudioRenderQuantum};
 use crate::SampleRate;
 
 use std::f32::consts::PI;
